Make code work w/ floyd-generated parser
This updates pyjson5 to use floyd to generate the parser instead
of glop, and to use an updated (floyd) grammar. Otherwise it attempts
to keep things as similar as possible (by generating the same AST, etc.).

We need two tiny changes in the tests because the floyd parser changes
the representation of unexpected characters in parse errors in a way
as to properly escape them.